fix seed script extension
